Moved search anywhere in entities to match convention
