+++ frontend/app/src/entities/search-anywhere/ui/search-anywhere-dialog.tsx
@@ -1,7 +1,8 @@
 import { Dialog, type DialogProps, Modal, ModalOverlay } from "react-aria-components";
 
-import { useSearchAnywhereContext } from "@/shared/components/search/search-anywhere-context";
 import { classNames } from "@/shared/utils/common";
+
+import { useSearchAnywhereContext } from "@/entities/search-anywhere/ui/search-anywhere-context";
 
 export function SearchAnywhereDialog({ children, className, ...props }: DialogProps) {
   const { isOpen, setIsOpen } = useSearchAnywhereContext();
