+++ frontend/app/src/entities/search-anywhere/ui/search-anywhere-empty.tsx
@@ -3,8 +3,8 @@ import { useCommandState } from "cmdk";
 
 import { INFRAHUB_API_SERVER_URL } from "@/config/config";
 
-import { SearchAnywhereGroup } from "@/shared/components/search/search-anywhere-group";
-import { SearchAnywhereItem } from "@/shared/components/search/search-anywhere-item";
+import { SearchAnywhereGroup } from "@/entities/search-anywhere/ui/search-anywhere-group";
+import { SearchAnywhereItem } from "@/entities/search-anywhere/ui/search-anywhere-item";
 
 export function SearchAnywhereEmpty() {
   const count = useCommandState((state) => state.filtered.count);
