+++ frontend/app/src/entities/search-anywhere/ui/search-anywhere-trigger.tsx
